Fishing base skill to fish in an area fix
Cherry-picked from Project Epoch.

Fixed an issue wherein fishing skill requirements for areas were significantly higher than intended.

As a more detailed technical answer, lets look at the Bay of Storms in Azshara for example. It requires 330 fishing to be able to catch any fish there, and at 425 it becomes guaranteed that you won't catch junk. In the DB it lists 425 for that subzone. So it was accidentally requiring 425 to be able to fish there, with the guarantee being 520. This PR reduces the fetched DB value by 95 to correct this.