Bypass CheckCasterAuras if the caster is not a unit.
Cherry-picked from Project Epoch.

This should solve hunter traps failing when the hunter is under any form
of CC.

As it stands in TrinityCore at the moment, m_originalCaster seems to
point to the Player both for the hunter casts and the trap casts, such
that the trap's spell cast will be checked for the hunter's caster
auras. Changing this might require up cleaning up code in a lot of other
places, so we just do an *actually proper* check for a unit caster and
we bail if the caster is a game object (i.e. a trap).