Solve imp cast range and wonky movement.
Cherry-picked from Project Epoch.

* Uncap caster pet chase distance.

Should solve issues like the imp not casting its firebolt at 30 yards.

* Do not pass ChaseAngle for MoveChase for caster pets.

Previously it passed angle = 0 with tolerance = 2 PI, which actually
meant that if we had a 30y chase distance it would path to 30 yards in
FRONT of the target, instead of straight to the target.

---------

Co-authored-by: zana244 <zana244@users.noreply.github.com>