Properly retrieve pet to dismiss for spells with SPELL_ATTR1_DISMISS_PET.
Cherry-picked from Project Epoch.

For some reason, the previous GetCreature function was not returning a
valid pointer when a warlock would use Enslave Demon with a summoned
pet, leading to the warlock having a pet + a charm. Modified to use the
more restrictive GetPet (since we are expecting a Pet...)

Co-authored-by: zana244 <zana244@users.noreply.github.com>